Dynamically generate api lists a bit
