+++ src/resources.rs
@@ -1,5 +1,4 @@
 use diff::Diff;
-use k8s_openapi::{NamespaceResourceScope, apimachinery::pkg::apis::meta::v1::APIResource};
 use serde::{Deserialize, Serialize};
 use std::{
     collections::BTreeMap,
@@ -66,44 +65,6 @@ impl Spec for ControllerRevision {
         &()
     }
 }
-
-macro_rules! impl_resource {
-    ($r:ident, $scope:ident, $apiversion:expr, $group:expr, $kind:expr, $version:expr, $urlpathsegment:expr) => {
-        impl k8s_openapi::Resource for $r {
-            type Scope = $scope;
-
-            const API_VERSION: &'static str = $apiversion;
-            const GROUP: &'static str = $group;
-            const KIND: &'static str = $kind;
-            const VERSION: &'static str = $version;
-            const URL_PATH_SEGMENT: &'static str = $urlpathsegment;
-        }
-    };
-}
-
-impl_resource!(Pod, NamespaceResourceScope, "v1", "core", "Pod", "v1", "pods");
-// impl_resource!(Job, "JobList");
-impl_resource!(Deployment, NamespaceResourceScope, "apps/v1", "apps", "Deployment", "v1", "deployments");
-// impl_resource!(ReplicaSet, "ReplicaSetList");
-// impl_resource!(StatefulSet, "StatefulSetList");
-// impl_resource!(PersistentVolumeClaim, "PersistentVolumeClaimList");
-// impl_resource!(Node, "NodeList");
-
-macro_rules! impl_listable {
-    ($r:ident, $kind:expr) => {
-        impl k8s_openapi::ListableResource for $r {
-            const LIST_KIND: &'static str = $kind;
-        }
-    };
-}
-
-impl_listable!(Pod, "PodList");
-// impl_listable!(Job, "JobList");
-impl_listable!(Deployment, "DeploymentList");
-// impl_listable!(ReplicaSet, "ReplicaSetList");
-// impl_listable!(StatefulSet, "StatefulSetList");
-// impl_listable!(PersistentVolumeClaim, "PersistentVolumeClaimList");
-// impl_listable!(Node, "NodeList");
 
 #[derive(
     Clone, Debug, Default, PartialEq, Eq, Hash, PartialOrd, Ord, Serialize, Deserialize, Diff,
